feat: Add meta units filter to lineup page
This commit adds a new checkbox item to the lineup page that allows users to filter units based on whether they are meta units or not. When the "Tylko meta jednostki" checkbox is checked, only units with a value greater than 7 will be displayed. This feature improves the user experience by providing a more focused view of the lineup units.

Based on recent user commits and repository commits, it appears that there have been multiple refactorings and feature additions related to the autocompleter component, including lazy loading for images and improved autocompletion.